requirements.txt
# Core dependencies
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.1.0
torch-geometric>=2.4.0

# NLP and Embeddings
sentence-transformers>=2.2.2
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2

# Vector Search
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Graph Database
neo4j>=5.14.0

# API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
python-multipart>=0.0.6

# Optional: GPU support (uncomment if using CUDA)
# torch==2.1.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torch-geometric==2.4.0 -f https://data.pyg.org/whl/torch-2.1.0+cu118.html
